move data-grid enums to a separate options file
