+++ packages/web-components/fast-components/src/data-grid/data-grid.definition.ts
@@ -1,6 +1,6 @@
-import { WebComponentDefinition } from "@microsoft/fast-tooling/dist/data-utilities/web-component";
+import { GenerateHeaderOptions } from "@microsoft/fast-foundation/dist/esm/data-grid/data-grid.options";
 import { DataType } from "@microsoft/fast-tooling";
-import { GenerateHeaderOptions } from "@microsoft/fast-foundation";
+import { WebComponentDefinition } from "@microsoft/fast-tooling/dist/data-utilities/web-component";
 
 export const fastDataGridDefinition: WebComponentDefinition = {
     version: 1,
